pyproject.toml
[tool.poetry]
name = "finder_tests"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.9"
Babel = "^2.10.3"
dash = "^2.6.2"
dash-bootstrap-components = "^1.2.1"
pandas = "^1.5.0"
pandas-stubs = "^1.5.0"
plotly = "^5.10.0"
flywheel-sdk = "^16.12.3"
packaging = "^21.3"
fw-client = "^0.4.0"
seaborn = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
